Removed the userfeedback modal and the toggle from the admin panel, removed old decoupled footer and added the ability to add custom links to footer. Also updated ui package to use the one with the removed consent wrds. Folloing tickets included. UOT-137810, UOT-137809, UOT-138036,
