+++ backend/node_app/routes/gameChangerRouter.js
@@ -200,8 +200,6 @@ router.get('/appSettings/intelligentAnswers', appSettings.getIntelligentAnswersM
 router.post('/appSettings/intelligentAnswers', appSettings.setIntelligentAnswersMode);
 router.get('/appSettings/entitySearch', appSettings.getEntitySearchMode);
 router.post('/appSettings/entitySearch', appSettings.setEntitySearchMode);
-router.get('/appSettings/userFeedback', appSettings.getUserFeedbackMode);
-router.post('/appSettings/userFeedback', appSettings.toggleUserFeedbackMode);
 router.get('/appSettings/jiraFeedback', appSettings.getJiraFeedbackMode);
 router.post('/appSettings/jiraFeedback', appSettings.toggleJiraFeedbackMode);
 router.get('/appSettings/topicSearch', appSettings.getTopicSearchMode);
