+++ frontend/package.json
@@ -19,7 +19,7 @@
 		"@datapunt/matomo-tracker-react": "0.3.1",
 		"@date-io/date-fns": "1.3.13",
 		"@dod-advana/advana-jira-feedback": "0.4.15",
-		"@dod-advana/advana-platform-ui": "1.1.1",
+		"@dod-advana/advana-platform-ui": "1.1.3",
 		"@dod-advana/advana-side-nav": "0.0.17",
 		"@dod-advana/advana-tutorial-overlay": "0.0.5",
 		"@emotion/react": "11.7.1",
